refactor app components
