+++ api/src/config.py
@@ -3,21 +3,19 @@ from sqlite3 import Connection
 from typing import Final
 from dotenv import load_dotenv
 
-if not load_dotenv("../.env.dev"):
-    raise Exception("No envs are set")
-
-BOT_TOKEN = os.environ["BOT_TOKEN"]
-LAMBDA_URL = os.environ["LAMBDA_URL"]
 BUILD = os.environ.get("BUILD", "prod")
-
 HOST = os.environ.get("HOST", None)
 WEBHOOK_HOST = HOST
 
 # Dev build code
-if BUILD.lower().startswith("dev"):
+if BUILD == "dev":
+    if not load_dotenv("../.env.dev"):
+        raise Exception("No envs are set")
     TUNNEL_URL = os.environ["TUNNEL_URL"]
     WEBHOOK_HOST = TUNNEL_URL
-    
+
+BOT_TOKEN = os.environ["BOT_TOKEN"]
+LAMBDA_URL = os.environ["LAMBDA_URL"]
 
 WEBHOOK_PORT = 8080
 WEBHOOK_PATH = f"/bot/{BOT_TOKEN}"
@@ -25,6 +23,11 @@ WEBHOOK_PATH = f"/bot/{BOT_TOKEN}"
 WEBHOOK_URL = f"{WEBHOOK_HOST or ''}{WEBHOOK_PATH}"
 
 WEBHOOK_LISTEN = "0.0.0.0"
+
+
+DB_FILENAME = "users.db" if BUILD.startswith("prod") else "dev.db"
+ADMIN = int(os.environ["ADMIN"])
+TEST_USER = int(os.environ["TEST_USER"])
 
 MSGS: Final[dict] = {
     "start": "Hello, this bot reboots your aws instance\n"
@@ -34,11 +37,6 @@ MSGS: Final[dict] = {
     "/reboot - reboot AWS Lightsail instance\n"
     "/help - show this message",
 }
-
-DB_FILENAME = "users.db" if BUILD.startswith("prod") else "dev.db"
-ADMIN = int(os.environ["ADMIN"])
-TEST_USER = int(os.environ["TEST_USER"])
-
 
 
 BotConfig: dict = dict(token=BOT_TOKEN, threaded=False)
